Add list tests to the correct file
