+++ packages/editor/CodeMirror/markdown/markdownCommands.toggleList.test.ts
@@ -350,4 +350,52 @@ A block quote:
 			getExpectedCursorLocations(editor.state.doc.toString()),
 		);
 	});
+
+	it('should convert a nested bulleted list to an ordered list', async () => {
+		const initialDocText = [
+			'- Item 1',
+			'    - Sub-item 1',
+			'    - Sub-item 2',
+			'- Item 2',
+		].join('\n');
+
+		const expectedDocText = [
+			'1. Item 1',
+			'    1. Sub-item 1',
+			'    2. Sub-item 2',
+			'2. Item 2',
+		].join('\n');
+
+		const editor = await createTestEditor(
+			initialDocText,
+			EditorSelection.range(0, initialDocText.length),
+			['BulletList'],
+		);
+
+		toggleList(ListType.OrderedList)(editor);
+
+		expect(editor.state.doc.toString()).toBe(expectedDocText);
+	});
+
+	it('should convert a mixed nested list to a bulleted list', async () => {
+		const initialDocText = `1. Item 1
+			1. Sub-item 1
+			2. Sub-item 2
+		2. Item 2`;
+
+		const expectedDocText = `- Item 1
+			- Sub-item 1
+			- Sub-item 2
+		- Item 2`;
+
+		const editor = await createTestEditor(
+			initialDocText,
+			EditorSelection.range(0, initialDocText.length),
+			['OrderedList'],
+		);
+
+		toggleList(ListType.UnorderedList)(editor);
+
+		expect(editor.state.doc.toString()).toBe(expectedDocText);
+	});
 });
